refactor(agent)!: the agent's bus is `agent::drive`
The runtime's bus (rig-bus) is about to become `rig_agent::bus`, so the
agent's own bus — `AgentBus`, `Driven`, the per-poll driver lock and the
waker set, "whoever holds the driver drives" — takes the name of its
first sentence: `agent::drive`, with `agent/drive/loom_models.rs`. The
loom shim moves to the crate root (`src/sync.rs`) so both protocols share
one; the workspace's check-cfg comment and the loom job's comment follow.

Paths only: `git diff -M` shows the three renames at 100% similarity
apart from the re-pathed imports. No behaviour change.

Gate: fmt; clippy -p rig-agent --tests -D warnings; rustdoc -D warnings;
nextest -p rig-agent (608 passed); tests/core (282 passed); loom
-p rig-bus -p rig-agent (10 models); rig-agent on wasm32.
