+++ Cargo.toml
@@ -73,8 +73,8 @@ exclude = [
 ]
 
 [workspace.lints.rust]
-# `--cfg rig_loom` selects the model checker's sync shims (`bus/sync.rs`,
-# `agent/sync.rs`); a build flag, not a feature, and crate-specific so no
+# `--cfg rig_loom` selects the model checker's sync shims (rig-bus's
+# `sync.rs`, rig-agent's `sync.rs`); a build flag, not a feature, and crate-specific so no
 # dependency reacts to it (`event-listener` reacts to a bare `loom`).
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(rig_loom)'] }
