+++ crates/rig-agent/src/agent/mod.rs
@@ -102,14 +102,13 @@
 //! # }
 //! ```
 mod builder;
-pub(crate) mod bus;
 mod completion;
+pub(crate) mod drive;
 mod engine;
 pub mod hook;
 pub mod run;
 pub mod runner;
 mod streaming;
-pub(crate) mod sync;
 mod telemetry;
 mod tool;
 mod typed;
