+++ crates/rig-agent/src/lib.rs
@@ -71,6 +71,7 @@ pub(crate) use rig_core::json_utils;
 pub mod prelude;
 pub mod run;
 pub mod streaming;
+pub(crate) mod sync;
 #[cfg(any(test, feature = "test-utils"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "test-utils")))]
 pub mod test_utils;
