+++ crates/rig-agent/src/sync.rs
@@ -1,6 +1,6 @@
-//! The synchronisation primitives the agent's driver protocol is written
+//! The synchronisation primitives the crate's driver protocols are written
 //! against: `std`'s, or `loom`'s under `--cfg rig_loom`, so the same code
-//! is what the model checker explores (`agent/bus/loom_models.rs`).
+//! is what the model checker explores (`agent/drive/loom_models.rs`).
 
 #[cfg(rig_loom)]
 pub(crate) use loom::sync::{Mutex, MutexGuard};
